Implemented Anthropic Library
Anthropic's pypl page: https://pypi.org/project/anthropic/
